Update README.md
This project demonstrates the use of unsupervised learning techniques to cluster cryptocurrencies based on their price changes. The process of finding the optimal k value, clustering with K-means, and visualizing the results provides valuable insights into the data. Dimensionality reduction with PCA further simplifies the data, making it easier to analyze and interpret.

The impact of using fewer features was analyzed, showing how dimensionality reduction can help in clustering without significant loss of information. This case study highlights the practical application of machine learning techniques in the field of data science and their potential to uncover meaningful patterns in large datasets.